+++ components/new-questionnaire-form/form-control-element/multiple-choice-input.vue
@@ -1,4 +1,6 @@
 <script lang="ts" setup>
+import type { QuestionSettingsMultipleChoice } from '~/components/new-questionnaire-form/type';
+
 const props = defineProps<{
   modelValue: QuestionSettingsMultipleChoice;
 }>();
