+++ components/new-questionnaire-form/form-control-element/single-choice-input.vue
@@ -1,4 +1,6 @@
 <script lang="ts" setup>
+import type { QuestionSettingsSingleChoice } from '~/components/new-questionnaire-form/type';
+
 const props = defineProps<{
   modelValue: QuestionSettingsSingleChoice;
 }>();
