+++ components/new-questionnaire-form/form-control-element/text-long-input.vue
@@ -1,4 +1,6 @@
 <script lang="ts" setup>
+import type { QuestionSettingsTextLong } from '~/components/new-questionnaire-form/type';
+
 defineProps<{
   modelValue: QuestionSettingsTextLong;
 }>();
