+++ components/new-questionnaire-form/user-specifier-input.vue
@@ -1,5 +1,6 @@
 <script lang="ts" setup>
 import { useFetchTraqApi } from '@/composables/type-fetch/traq/use-fetch-traq-api';
+import type { UserSpecifier } from '~/components/new-questionnaire-form/type';
 
 const props = defineProps<{
   modelValue: UserSpecifier;
